Fix row visibility when calling create_key() from with an INSERT
The create_key() function is calling itself to create a new
parent_key when inserting a raw_key without existing parent_key
to encrypt it.
During the encryption trigger call, the newly created key was not
visible from SPI query to gather its meta datas, even though the
parent_key command is finished before the trigger fires.

This visibility mechanism is explained in this page:
https://www.postgresql.org/docs/current/spi-visibility.html