Implement generic encryption trigger in C
This commit creates C trigger functions
"trg_encrypt_using_key_col()" and "trg_encrypt_using_key_id()"
able to encrypt any column in any table no matter the name or
position of the column.

That means we now pass to the trigger the column name we need
to encrypt and other infos about the encryption key, additional
data, nonce, etc.

In consequence, we now use the same two functions for all our
encryption triggers. This avoids the generation of a dedicated
plpgsql trigger for each column we need to encrypt.

Creating generic triggers for any table/column is not possible in
plpgsql because you can not reference a record's field using a
variable. Plpgsql is not able to dereference a column inside the
"new" record.